Fixes asdf poetry & python upgrade action (#43)
## Description

Python & Poetry upgrades have not been working for a longer time. This
PR fix the issue by:

1. Updating `asdf-vm/actions/plugins-add` to the newest version. Old SHA
was pointing to the version which was already ~2 years old.
2. Using `poetry lock` instead of `poetry lock --no-update` for poetry
version 2+ as nowadays poetry defaults to no-update on `poetry lock` and
support for that flag was removed.
3. Removing trailing 't' from python version returned by `asdf latest
python` (experimental free threading). `asdf latest python` returns for
example `3.13.3t` instead of 3.13.3 for python 3.13+. t This causes
installation to fail due missing packages. At this point we don’t need
to install variant of experimental support for free threading.

Additionally, the PR adds also `major` to the `version` options for
visibility. It was missing from the listing even though the code assumed
it to be one of the options.


## Checklist
- [ ] After merging this I have tested this by manually triggering the
action in one of the internal projects